apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
statsmodels==0.14.5
faiss-cpu==1.12.0
pypdf==5.9.0
pytest==8.4.1
httpx==0.28.1
APScheduler==3.11.0
reportlab==4.2.5
